+++ src/main/java/io/hardingadonis/saledock/dao/impl/OrderDAOImpl.java
@@ -43,8 +43,8 @@ public class OrderDAOImpl implements IOrderDAO {
     }
 
     @Override
-    public Long count() {
-        Long count = 0L;
+    public Integer count() {
+        Integer count = 0;
 
         try {
             Connection conn = Singleton.dbContext.getConnection();
@@ -54,7 +54,7 @@ public class OrderDAOImpl implements IOrderDAO {
             ResultSet rs = smt.executeQuery();
 
             if (rs.next()) {
-                count = rs.getLong(1);
+                count = rs.getInt(1);
             }
 
             Singleton.dbContext.closeConnection(conn);
@@ -66,8 +66,8 @@ public class OrderDAOImpl implements IOrderDAO {
     }
 
     @Override
-    public Long countOrderInProcess() {
-        Long count = 0L;
+    public Integer countOrderInProcess() {
+        Integer count = 0;
 
         try {
             Connection conn = Singleton.dbContext.getConnection();
@@ -77,7 +77,7 @@ public class OrderDAOImpl implements IOrderDAO {
             ResultSet rs = smt.executeQuery();
 
             if (rs.next()) {
-                count = rs.getLong(1);
+                count = rs.getInt(1);
             }
 
             Singleton.dbContext.closeConnection(conn);
